Remove doc-lsp module (moved to connectroot/developertools)
Inline hashContent into sources.ts, remove chokidar/gray-matter/remark-frontmatter
dependencies, delete doc-lsp tests and hash-consistency tests.
