+++ package.json
@@ -43,12 +43,8 @@
     "@dagrejs/graphlib": "^4.0.1",
     "@inquirer/prompts": "^8.3.2",
     "@modelcontextprotocol/sdk": "^1.0.0",
-    "chokidar": "^5.0.0",
     "commander": "^14.0.3",
-    "gray-matter": "^4.0.3",
     "js-yaml": "^4.1.0",
-    "remark": "^15.0.1",
-    "remark-frontmatter": "^5.0.0",
     "remark-parse": "^11.0.0",
     "unified": "^11.0.5"
   },
