+++ src/sources.ts
@@ -8,13 +8,28 @@
  * hashing.
  */
 
+import crypto from "node:crypto";
 import fs from "node:fs";
 import path from "node:path";
 import type { GraphDefinition } from "./schema/graph-schema.js";
-import { hashContent } from "./doc-lsp/hash.js";
 
-// Re-export for consumers
-export { hashContent } from "./doc-lsp/hash.js";
+// --- Content hashing ---
+
+function normalizeContent(content: string): string {
+  return content.replace(/\r\n/g, "\n").trimEnd();
+}
+
+/**
+ * Hash content using SHA-256, returning the first 16 hex characters.
+ * Content is normalized (CRLF→LF, trimEnd) before hashing.
+ */
+export function hashContent(content: string): string {
+  return crypto
+    .createHash("sha256")
+    .update(normalizeContent(content))
+    .digest("hex")
+    .substring(0, 16);
+}
 
 // --- Types ---
